LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/filters/2d.filter.canvasFilterObject.blur.exceptions.tentative-expected.txt

2d.filter.canvasFilterObject.blur.exceptions.tentative

Test exceptions on gaussianBlur filter


FAIL Test exceptions on gaussianBlur filter assert_throws_js: function "function() { ctx.filter = new CanvasFilter(
    {name: 'gaussianBlur'}); }" threw object "ReferenceError: Can't find variable: CanvasFilter" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
